+++ src/XUnity.AutoTranslator.Plugin.Core/Properties/AssemblyInfo.cs
@@ -1,0 +1,7 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Runtime.CompilerServices;
+using System.Text;
+
+[assembly: InternalsVisibleTo( "XUnity.AutoTranslator.Plugin.Core.Tests" )]